opensearch_service/requirements.txt
astapi
uvicorn
pydantic
requests
elasticsearch[async]==8.12.0  # Azure OpenSearch compatible
azure-search-documents


